update y delete de marca

+++ src/main/java/pe/edu/upc/entities/Tienda.java
@@ -43,7 +43,7 @@ public class Tienda {
 	@Min(value = 100000000, message = "El telefono debe tener 9 digitos")
     @Max(value = 999999999, message = "El telefono no debe ser mayor a 9 digitos")
 	@Column(name = "telefonoTienda", nullable = false)
-	private double telefonoTienda;
+	private int telefonoTienda;
 	
 	@Column(name = "webTienda",nullable = false)
 	private String webTienda;
